Make the expanded mock data match the compact data (#11)
Heads-up @alexhebing: this should hopefully answer your concern in #25.
The @type of a node turns out to look indeed like this

@type: ['IRI', 'IRI']

and not like this, as we previously thought:

@type: [{
    @id: 'IRI',
}, {
    @id: 'IRI',
}]

so you were definitely right to raise that issue. You'll need to
update your code, as well as any mock data that live outside of the
mock-data directory.