Simplify context management considerably (#96)
The old context logic was based on the assumption that the
processContext function from the jsonld package produces a "pure",
fully resolved JSON-LD context object. This turned out not to be true;
it resolves to an internal specialized datastructure, which is
basically an undocumented implementation detail. This necessitated
our adjustment in 2572e28a. While this solved the error message we
were facing at the time, the unmet assumption still meant that the
logic was essentially invalid. This commit solves the problem forever
by not trying to resolve the context ahead of time.

@alexhebing This is the result of our discussion on a very warm day.