Add the first syncLD test (#96)
This test fails for reasons that will be addressed in subsequent
commits.

Successful compilation depends on the changes in
https://github.com/DefinitelyTyped/DefinitelyTyped/pull/37804.
The above pull request hasn't been merged yet at the time of writing.
In order to run the tests in the meanwhile, patch your local
frontend/node_modules/@types/jasmine-ajax/index.d.ts with the changes
in the pull request.