Define single source of truth for common prefix of our namespaces (#83)
The acute reader may notice that this is done independently in the
backend, in readit/settings.py, and the in the frontend, in the
config.json (which is then overridden in the project root). This may
seem like repetition, but it is not; it is conceivable that the
frontend and the backend will be running on separate servers.