Define a fixture module producing the canonical graph contents (#83)
@alexhebing As a starting point, this is just a copy of the mock-vocab
in our frontend, except that I left out the creator/created metadata.
I don't think those properties will be very applicable to any of our
graphs, except for the items.

I considered saving the fixture as a JSON-LD document. I decided
against it, because a (maintainable) JSON-LD document depends on a
context. It is too early to say with certainty what that context
should look like, so we'd likely end up changing the context multiple
times and having to edit the fixture every time. By writing the
fixture in rdflib notation in Python, I get a maintainable shorthand
notation that doesn't depend on any context. It also saves some file
operations when using the fixture.