Initial version of the item summary block (#38).
@jgonggrijp: a question. This block now counts on ontology instances to have a label (see the previous commit for examples).
For example, we have a 'Content' (say the book 'Jane Eyre'). Will 'Jane Eyre' be its label?
Or is it a title? Will this value exist in two fields?

Consider a Reader: I think we want to display it's name here. It will be in a field like
foaf:Name (or whatever), so how do we display it here? Make sure skos.prefLabel is also set when
the instance is created?