Avoid highlights in whitespace to prevent empty rects in Safari (#122)
@alexhebing Note that the first text node of the createSomeHtml
template starts with a newline character and 16 spaces. This becomes
invisible once inserted into the DOM, but the text is still in the
node (Chrome and Safari actually agree about this). When aRange
contains only this invisible whitespace, aRange.getClientRects()
produces an empty list in Safari but a list with a single rect that
takes up no space in Chrome. I don't know which one is right
according to the standards, but I do know that it is more realisting
to select printing characters, so that's what I did by changing the
offsets in this commit.

Sums are for clarification only. In the next commit, I replace them
by hard numbers.