Fix broken overlapdetails (#194).
@jgonggrijp: FYI: taking highlight views out of the DOM (i.e. detach) BREAKS the Range object
that they were carrying / referencing, in the sense that the Range doesn't have any indices (position details)
any more, and thus no content.

These details will always be stored in the view's AnnotationPositionDetails (so no big deal in that sense),
but the Range is needed to extract the view's text content. I worked around this by storing the text immediately when
first loading the Range.

Just a heads up about the ever-stranger Range object :confused: