Remove an unexpected special case from getLabelFromId utility (#66, #123)
@alexhebing FYI. I felt justified in removing these lines because (1)
the old behaviour was rather unexpected for me as an "outsider" and
(2) there didn't appear to be any code that relied on it, at least
not anymore. If such code never existed in the first place, please
learn from this that it is better to omit special cases from your
code, unless there is an external reason to include them.