Change allauth config to allow users without verified email to login.
@jgonggrijp: if you have a branch based on develop that has the old settings,
please merge this commit. In my case, it was quite a hassle to change these
settings after attempting to login via the frontend with an 'old' account.
The hassle, fyi, had to do with Django crashing, which resulted in multiple gulp
instances, Django not wanting to start because it was already running (although
it wasn't), etc. In other words: kill kill kill!