Add a test to check popUntilAndPush against premature push (#279)
It passes.

I wanted to test with a 1000 panels, because @alexhebing's infinite
loop detection triggers at 999 panels and I wanted to rule out that
this is playing a role in any bug.

I was continuously measuring the time taken by this test while
implementing the optimizations in the previous commits. Initially,
just pushing 1000 empty panels took nearly 90 seconds because every
push was accompanied by a binary search. Removing the binary search
container reduced this to about 1.5 seconds, but the test still timed
out if I next ran the `popUntilAndPush` that the test was actually
about.

I temporarily lowered the number of panels to 100 so I could measure
the total time taken without the test timing out. The fastTimeout
optimization sped up iteration with another factor 10, so that real
browsers can now run the entire test with 1000 panels within a
second. Unfortunately, JSDOM simply cannot offer this performance, so
the unittest is stuck at 100 panels when run from the console.

@alexhebing FYI.