Go for the normalize option after all (#283, #285)
The two alternative solutions turned out to be not entirely
equivalent. The old method of inserting the text with a template
string treats \r\n as a single character while .text() treats it as
two. This may cause pre-existing highlights to appear misaligned if
the source contains CRLF sequences. The former method is also more in
agreement with Python's way of decoding to str. I did however discover
that the template string approach trims leading newlines and possibly
other whitespace.

@JeltevanBoheemen FYI.