Process final review by @BeritJanssen and @JeltevanBoheemen (#412)
As discussed. I decided not to make the item/X references clickable
for now, because the item/X route itself is likely not be represented
by a panel. If I were to insert a clickable link to that route anyway,
a curious user may be unpleasantly surprised by the entire explorer
suddenly being reset.

Of course, I could do something more sophisticated, but that would be
more time-consuming. At least, this serves the purpose of clarifying
what the user is seeing.