Use select2 in AnnotationEditPanel, for real now (#403 #419)
@BeritJanssen you were right about importing select2 as a browserLib.
The CommonJS variant of selec2, which would be used otherwise, is
broken.

I removed the className of itemPicker so that the Bulma styling would
not interfere with with select2 styling. The way that select2 is
currently injected into the itemPicker from the outside is very hacky,
this should be refactored.