Do actually replace ItemGraph contents as advertized #163 #501 #253 #436
The remove: false option was added in c100042d in order to enable
paginated annotation fetching (#253). We however removed that feature
again in da75f42f (hotfix 0.11.1). It was not being used anywhere else
and now, it caused the userItems and userSources globals to wrongly
retain the user's items and sources after logout (and even across a
subsequent login) (#163). If we ever need to retain nodes on a
subsequent query again, we should instead opt to pass an aditional
options argument.