Use question marks instead of zeros as initial tallies on landing
We were seeing significant delays until the actual numbers came in,
due to an issue in the Fuseki triplestore. Seeing a "you made 0 items"
message might be alarming for users who did actually make items, so
the question marks are there to convey that the data are still being
retrieved.