Restore file change handler in frontend SourceUploadForm (#37 #517)
The render(Container) method was never the right place to bind events
to child elements, but we got away with it when the form was still a
monolithic view. In the new composite situation, the element in
question does not exist yet during the call to renderContainer. The
conventional delegate mechanism completely avoids this problem.